Add Scanner type.

+++ src/token.rs
@@ -5,11 +5,11 @@ use crate::token_type::TokenType;
 pub struct Token {
     token_type: TokenType,
     lexeme: String,
-    line: u32
+    line: usize
 }
 
 impl Token {
-    pub fn new(token_type: TokenType, lexeme: String, line: u32) -> Token {
+    pub fn new(token_type: TokenType, lexeme: String, line: usize) -> Token {
         Token { token_type, lexeme, line }
     }
 }
